+++ src/view/home/bottom_bar.rs
@@ -45,8 +45,8 @@ impl BottomBar {
 
         let page_label = PageLabel::new(rect![pt!(rect.max.x - side - big_half_width, rect.min.y),
                                               pt!(rect.max.x - side, rect.max.y)],
-                                        current_page as f32,
-                                        pages_count as f32,
+                                        current_page as f64,
+                                        pages_count as f64,
                                         false);
         children.push(Box::new(page_label) as Box<View>);
 
@@ -77,7 +77,7 @@ impl BottomBar {
 
     pub fn update_page_label(&mut self, current_page: usize, pages_count: usize, hub: &Hub) {
         let page_label = self.children[2].as_mut().downcast_mut::<PageLabel>().unwrap();
-        page_label.update(current_page as f32, pages_count as f32, hub);
+        page_label.update(current_page as f64, pages_count as f64, hub);
     }
 
     pub fn update_icons(&mut self, current_page: usize, pages_count: usize, hub: &Hub) {
